refactor: command 패키지 위치 이동
